Settings: Allow customizing default notification vibration pattern [2/2]
Re-use the patterns we've got for ringtone vibration as options
In addition allow crafting a custom pattern same as we do for ringtone

Re-use as much resources and controllers as possible from our ringtone vibration pattern options.
To achieve that make the original controllers more generic and inherit in our notification ones.

Change-Id: If8de200ed7c4e2b16b0d88e2129062248751545c